Implemented ResourceLoader and Test
Added new constructors for ‘resources’ (textures, shaders, etc). Takes
a file path and creates the resource on the spot. For shaders, don’t
add an extension. It will search for all shaders that have the same
name and create the program on them. I might change this later to take
in a custom shader file format.